requirements.txt
matplotlib==3.5.2
mlxtend==0.20.0
numpy==1.21.5
pandas==1.4.2
Pillow==9.2.0
seaborn==0.11.2
streamlit==1.10.0
